refactor(pipeline): rename RouterCreator to Pipeline, move it into the pipeline module
RouterCreator is the product of pipeline construction, not a factory:
the struct create_pipeline returns and the state machine holds across
reloads, carrying the router service, the plugins' web endpoints, the
pipeline handle, and the seeds the next reload inherits (configuration
and the in-memory query-plan cache). Name it Pipeline and house it,
its constructor, and its RouterFactory impl in pipeline/mod.rs, so the
router stack assembly sits with the other construction-time stacks.
RouterService and the request path stay in services/router/service.rs.
