+++ apollo-router/src/services/router/tests.rs
@@ -16,12 +16,12 @@ use tower_service::Service;
 
 use crate::Context;
 use crate::graphql;
+use crate::pipeline::from_supergraph_mock;
 use crate::plugin::test::assert_no_mock_calls;
 use crate::plugin::test::await_mock_driver;
 use crate::services::SupergraphRequest;
 use crate::services::SupergraphResponse;
 use crate::services::router;
-use crate::services::router::service::from_supergraph_mock;
 use crate::services::router::service::process_vary_header;
 use crate::services::subgraph;
 use crate::services::supergraph;
